moved material creation and all of material_loading into material_creation.cpp so that no one ever needs to know about the material definition struct unless they want to create a new mat

+++ VkMaterialSystem/hash.cpp
@@ -5,13 +5,13 @@ uint64_t MurmurHash64A(const void * key, int len, unsigned int seed);
 
 uint32_t hash(const char* str)
 {
-	return MurmurHash2(str, strlen(str), 255);
+	return MurmurHash2(str, static_cast<int>(strlen(str)), 255);
 }
 
 
 uint64_t hash64(const char* str)
 {
-	return MurmurHash64A(str, strlen(str), 255);
+	return MurmurHash64A(str, static_cast<int>(strlen(str)), 255);
 }
 
 //-----------------------------------------------------------------------------
